+++ src/main/java/com/khem/appspring/springphoneshop/repository/ColorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.khem.appspring.springphoneshop.model.Color;
 
-public interface ColorRepository extends JpaRepository<Color,Integer>{
+public interface ColorRepository extends JpaRepository<Color,Long>{
     
 }
